refactor link pattern into corresponding models
