+++ _1327/documents/markdown_internal_link_pattern.py
@@ -1,0 +1,14 @@
+import markdown
+from markdown.inlinepatterns import LinkPattern
+
+
+class InternalLinkPattern (LinkPattern):
+
+	def handleMatch(self, m):
+		el = markdown.util.etree.Element("a")
+		el.set('href', self.url(m.group('id')))
+		el.text = markdown.util.AtomicString(m.group('title'))
+		return el
+
+	def url(id):
+		raise NotImplementedError
