+++ _1327/information_pages/models.py
@@ -2,6 +2,7 @@ from django.core.urlresolvers import reverse
 from django.utils.translation import ugettext_lazy as _
 from reversion import revisions
 
+from _1327.documents.markdown_internal_link_pattern import InternalLinkPattern
 from _1327.documents.models import Document
 
 
@@ -18,6 +19,14 @@ class InformationDocument(Document):
 		permissions = (
 			(INFORMATIONDOCUMENT_VIEW_PERMISSION_NAME, 'User/Group is allowed to view that document'),
 		)
+
+	class LinkPattern (InternalLinkPattern):
+
+		def url(self, id):
+			document = InformationDocument.objects.get(id=id)
+			if document:
+				return reverse('information_pages:view_information', args=[document.url_title])
+			return ''
 
 	def get_view_url(self):
 		return reverse('information_pages:view_information', args=(self.url_title, ))
