+++ _1327/information_pages/views.py
@@ -15,7 +15,7 @@ from markdown.extensions.toc import TocExtension
 from reversion import revisions
 
 from _1327.documents.forms import PermissionForm
-from _1327.documents.markdown_internal_links import InternalLinksMarkdownExtension
+from _1327.documents.markdown_internal_link_extension import InternalLinksMarkdownExtension
 from _1327.documents.models import Document
 from _1327.documents.utils import (
 	delete_old_empty_pages,
